# team_task_hub/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.SendVerificationCodeRequest:
    properties:
      business:
        description: '业务类型: register, reset_password, change_email, join_org'
        type: string
      email:
        description: 邮箱地址
        type: string
    required:
    - business
    - email
    type: object
...
  handlers.VerifyCodeRequest:
    properties:
      business:
        description: 业务类型
        type: string
      code:
        description: 6位验证码
        type: string
      email:
        description: 邮箱地址
        type: string
    required:
    - business
    - code
    - email
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 团队任务管理平台API
  title: Team Task Hub API
  version: "1.0"
paths:
  /api/admin/email/cleanup:
    post:
      description: 清理数据库中已过期的验证码记录（定时任务调用）
      produces:
      - application/json
      responses:
        "200":
          description: 清理完成
          schema:
            type: string
        "500":
          description: 清理失败
          schema:
            type: string
      summary: 清理过期验证码
      tags:
      - 系统管理
...
  /api/admin/email/test-connection:
    get:
      description: 测试邮箱SMTP服务器连接状态（仅用于调试）
      produces:
      - application/json
      responses:
        "200":
          description: SMTP连接正常
          schema:
            type: string
        "500":
          description: SMTP连接失败
          schema:
            type: string
      summary: 测试SMTP连接
      tags:
      - 系统管理
...
  /api/email/send-verification:
    post:
      consumes:
      - application/json
      description: 向指定邮箱发送6位数字验证码，支持多种业务场景，"register","change_email_old","change_email_new","change_password"
      parameters:
      - description: 发送验证码请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendVerificationCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 验证码发送成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            type: string
        "429":
          description: 请求过于频繁
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      summary: 发送邮箱验证码
      tags:
      - 邮箱验证
  /api/email/verify:
    post:
      consumes:
      - application/json
      description: 验证邮箱和验证码是否匹配且有效，支持多种业务场景
      parameters:
      - description: 验证验证码请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 验证码验证成功
          schema:
            type: string
        "400":
          description: 验证码无效或已过期
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      summary: 验证邮箱验证码
      tags:
      - 邮箱验证
...
swagger: "2.0"